rename atlasmapWebView to atlasMapWebView #36
and it automatically modified the import by removing the useless ./
